#1636 stop using deprecated kubelet flags and move it to the config
